requirements.txt
alembic==1.11.1
aniso8601==9.0.1
apispec==6.3.0
async-timeout==4.0.2
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.9.0
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
Flask==2.3.2
flask-apispec==0.11.4
Flask-Caching==2.0.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
pluggy==1.0.0
psycopg2
pyasn1==0.5.0
pytest==7.3.1
pytest-flask==1.2.0
pytest-flask-sqlalchemy==1.1.0
pytest-mock==3.10.0
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
redis==4.5.5
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.15
tomli==2.0.1
typing_extensions==4.6.2
webargs==8.2.0
Werkzeug==2.3.4
WTForms==3.0.1
zipp==3.15.0
